refactor(features/partyroom): useRegisterMeToQueue mutation 승격 (widget → feature)
데스크탑 widget cross-import 회피 + 모바일 chunk 4 의 features-mobile 합성 hook 에서 import 가능.
동작 0 변경 (mutation 본문 1:1 복사, import path 만 갱신).
