+++ src/features/partyroom/register-me-to-queue/index.ts
@@ -1,0 +1,1 @@
+export { useRegisterMeToQueue } from './api/use-register-me-to-queue.mutation';
